Cleanup auth helpers
