+++ src/features/auth/helpers.ts
@@ -1,0 +1,40 @@
+import Cookies from 'js-cookie';
+
+import { setAuthToken } from './token';
+import { saveAllData, emptyData, IAuth } from './useSavedAuth';
+
+export const AUTH_COOKIE = 'coordinape_auth_cookie';
+
+export const clearSavedAuth = () => saveAllData(emptyData());
+
+export const logoutAndClearSavedAuth = () => {
+  clearSavedAuth();
+  setAuthToken('');
+};
+
+export const setAuthTokenForAddress = (address: string, token: string) => {
+  const allData = getAllData();
+  allData.data[address.toLowerCase()] = { token };
+  allData.recent = address.toLowerCase();
+  saveAllData(allData);
+  setAuthToken(token);
+};
+
+export const reloadAuthFromCookie = () => {
+  const allData = getAllData();
+  const { recent, data } = allData;
+  if (recent) {
+    const auth = data[recent] ?? {};
+    setAuthToken(auth.token);
+    return true;
+  }
+};
+export const getAllData = (): IAuth => {
+  try {
+    const stored = Cookies.get(AUTH_COOKIE);
+    if (!stored) return emptyData();
+    return JSON.parse(stored);
+  } catch {
+    return emptyData();
+  }
+};
